#21.1000/feature/sannikov/research/4 Декомпозиция методов рендеренга https://online.sbis.ru/opendoc.html?guid=600c8026-e058-42d4-8bb8-8365c18cae44
